+++ django_tables2_column_shifter/tests/settings.py
@@ -42,3 +42,5 @@ MEDIA_URL = '/media/'
 STATIC_URL = '/static/'
 
 MIDDLEWARE = []
+
+DEFAULT_AUTO_FIELD = 'django.db.models.AutoField'