+++ testproject/testproject/tables.py
@@ -7,6 +7,8 @@ def get_author_table_class(TableClass):
 
     class AuthorTable(TableClass):
 
+        column_excluded = ['id']
+
         class Meta:
             model = Author
             attrs = {'class': 'table table-bordered table-striped table-condensed'}
@@ -18,6 +20,7 @@ def get_book_table_class(TableClass):
 
     class BookTable(TableClass):
 
+        column_excluded = ['id']
         column_default_show = ['id', 'title', 'author']
 
         class Meta: